Refactor language configuration out of Next.js page
